Removed civic_barbaric_despoilers_corporate in favor of civic_barbaric_despoilers
The modifications to the "slave" pop category were causing crashes in
early mid-game. The only difference between the mod's version of the pop
category and the vanilla version is the addition of code handling the
corporate version of the Barbaric Despoilers civic. However, the
corporate and non-corporate versions of the civic were identical, except
for the corporate version's requirement to have the Pirate Lord civic.

So, the non-corporate version of the civic was changed to be available
for either the normal vanilla conditions OR the presence of the Pirate
Lord civic. This allowed the removal of the corporate version of the
civic and the removal of the modified version of the "slave" pop
category, fixing the crash.