Fix for human empires not avalible
the human empires need the humanoids Species Pack to be playable properly, this is a fix that makes identical alt versions that don't use the DLC ship and city sets and with a tag to only show up when the DLC is *not* enabled. This is how they handle it in vanilla code too.